Update LipidCA_main.m
Run without the need for the computeAreas.m function, which is missing. @mickjt forgot to push this function. I created an issue to let him know about this. Nevertheless, it seems like working without this function. That is the reason I commented out the lines where computeAreas.m is called.